pm-agent/pyproject.toml
[tool.poetry]
name = "pm-agent"
version = "0.1.0"
description = "Personal Jira PM agent built on Google ADK + VertexAI"
authors = ["Jackson McKissack"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
google-adk = ">=1.0.0"
google-cloud-aiplatform = {version = ">=1.112", extras = ["agent_engines", "adk"]}
google-genai = "*"
requests = ">=2.31"
python-dotenv = ">=1.0"
pydantic = ">=2.0"
pydantic-settings = ">=2.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.23"
ruff = ">=0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
